requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.6
dateparser==1.0.0
defusedxml==0.7.0rc2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.10
django-admin-sortable2==0.7.8
django-analytical==3.0.0
django-ckeditor-5==0.0.12
django-classy-tags==2.0.0
django-cms==3.7.4
django-colorfield==0.3.2
django-filer==2.0.2
django-fontawesome-5==1.0.18
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-polymorphic==3.0.0
django-recaptcha==2.0.6
django-sekizai==2.0.0
django-summernote==0.8.11.6
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
easy-thumbnails==2.7
filelock==3.0.12
html5lib==1.1
idna==2.10
isort==5.4.2
jet-bridge-base==0.9.3
jet-django==1.0.6
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.0
pipenv==2020.8.13
prompt-toolkit==2.0.9
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==4.0.3
SQLAlchemy==1.3.23
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
